pyproject.toml
[tool.poetry]
name = "simple-echo-bot"
version = "0.1.0"
description = "A simple echo IM bot based on nonebot2"
authors = [ "Luo Chen <luochen1990@gmail.com>" ]
include = ["src"]

[tool.poetry.dependencies]
python = ">=3.10,<4"
nonebot2 = {extras = ["fastapi", "httpx"], version = "^2.3.2"}
nb-cli = "^1.4.1"
nonebot-adapter-telegram = "^0.1.0b17"
nonebot-adapter-qq = "^1.4.4"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-xdist = "^3.6.1"
pytest-cov = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.373"
funix = "^0.5.8"

[tool.pytest.ini_options]
testpaths = ["test"]
addopts = "-n auto"

[tool.nonebot]
adapters = [
    { name = "Telegram", module_name = "nonebot.adapters.telegram" },
    { name = "QQ", module_name = "nonebot.adapters.qq" },
]
plugins = []
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
